// src/auth/enums/permission.enum.ts
export enum Permission {
  JOURNAL_CREATE = 'journal:create',
  JOURNAL_READ = 'journal:read',
  JOURNAL_UPDATE = 'journal:update',
  JOURNAL_DELETE = 'journal:delete',

  POEM_READ = 'poem:read',
  POEM_PUBLISH = 'poem:publish',
  POEM_EDIT = 'poem:edit',
  POEM_SHARE = 'poem:share',

  USER_MANAGE = 'user:manage',
}